[Refactor]: City 위도,경도 double 타입으로 재변경

+++ src/main/java/com/minair/domain/City.java
@@ -23,11 +23,11 @@ public class City {
     private Long id;
 
     private String name;
-    private float longitude;
-    private float latitude;
+    private double longitude;
+    private double latitude;
 
     @Builder
-    public City(String name, float longitude, float latitude) {
+    public City(String name, double longitude, double latitude) {
         this.name = name;
         this.longitude = longitude;
         this.latitude = latitude;
